сущности в отдельный файл
